Fixed import errors from file reorganization, added skeleton User Annotation endpoint
