+++ server/api/routes/analysis/inputs.js
@@ -5,21 +5,21 @@ const multer = require('multer');
 const mkdirp = require('mkdirp');
 const config = require('config');
 
-const logger = require('../../util/logger');
+const logger = require('../../../util/logger');
 
-const { arrayDiff } = require('../../util/array');
-const { getUploadPath } = require('../../util/storage');
-const { getAudioMetadata } = require('../../util/audioMetadata');
+const { arrayDiff } = require('../../../util/array');
+const { getUploadPath } = require('../../../util/storage');
+const { getAudioMetadata } = require('../../../util/audioMetadata');
 
-const DAinputs = require('../../data_access/inputs/DAinputs');
+const DAinputs = require('../../../data_access/inputs/DAinputs');
 
 const {
   parseInputJson,
   getNewFilename,
   getExtensionlessFilename,
-} = require('../models/input/utils');
+} = require('../../models/input/utils');
 
-const { verify } = require('../../util/verify');
+const { verify } = require('../../../util/verify');
 
 const error = config.get('error');
